V.0.0.2: Major Order & War State Parsers Working
We finally have proper fetchers and parsers working. I was able to slowly implement a correct approach to calculate the expiration date; the API delivers an integer in seconds, so a function needed to be created to read both the current date and the remaining seconds to calculate the end time.